pyproject.toml
[tool.poetry]
name = "ict-4007"
version = "0.1.0"
description = ""
authors = ["Ian Kirkpatrick"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
loguru = "^0.7.3"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
